use completion in api client
